function stratified_sampling moved to preprocess\utils
